Weapon Fix
Weapons do not destroy themselves at 0 Attack. Instead, a Hero is unable to attack (As they have 0 Attack, much like minions.) It is possible to have 0 or less Attack on a Weapon, but 0 is always displayed for non-positive Attack values.

Also, Weapon buffs and Attack buffs are exclusively applied, so negative Attack Weapons do NOT affect Hero Attack buffs.