Fixed Prince Malchezaar
This card is kind of a pain in the butt to make. Currently, this card
does nothing in Sandbox, as the necessary event is ignored initially
(This is necessary to dodge the mulligan phase.) I consider this to be
an unintended bug that goes along with the card, since the card itself
works in Play Mode. In order to make it work in Sandbox Mode, events
need to be set to ignore AFTER the GameStartEvent, which probably won't
happen. The biggest change this makes, however... Is that you can now
have cards that interact with the game... from your deck. A card that is
IN YOUR DECK can put a trigger into effect. The issue is that this
currently only works for cards that you start the game with. Any card
that is shuffled in later will never add a deck trigger. Since this
seems to be a unique type of card, and having a card that interacts from
your deck can be incredibly broken, I'm leaving it as is. If someone
wants to mess about with it, be my guest, but I'm not extending it until
I see a good reason to.