Within group 4 will be seen the hero, hero power, basic and classic cards of the custom class "Wraith". The class colour as can be seen in the wraith folder; is crimson (including the RGB values)
I have discussed with Doc, the new keyword "pacts", which he said he'd do the code for (assuming its not him reading). they are clearly labelled in the classic set file and have been coded like quests for now. The only change needed *should* be where it says "quest", instead say "pact".

I also have given skeleton tokens in the class, a new attribute as "SKELETON", however if there is coding needed beyond saying they have the attribute in their own code: I have not done that, sorry. And that will need to be done because of future cards that synergise with skeletons in the same way treant synergies work.

Sincerely, Bromora.